Moved Welcome baskets into HTTPBuildableDirectoryMap/DiscoverableDirectories namespace.
